+++ app/(tabs)/(more)/index.tsx
@@ -1,44 +1,25 @@
 import React from 'react';
-import { Pressable, StyleSheet, Text } from 'react-native';
+import { StyleSheet, Text } from 'react-native';
 
 import { router } from 'expo-router';
 
 import { ThemedButton } from '@/components/ThemedButton';
-import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
-import { useColors } from '@/hooks/useColors';
 
 export default function MoreScreen() {
-  const { primaryColor } = useColors();
   return (
     <ThemedView style={styles.container}>
-      <ThemedButton
-        onPress={() => router.push('/settings')}
-        style={[styles.button]}
-        variant="primary"
-      >
-        <Text style={styles.buttonText}>الإعدادات</Text>
+      <ThemedButton onPress={() => router.push('/settings')} variant="primary">
+        <Text>الإعدادات</Text>
       </ThemedButton>
-      <ThemedButton
-        onPress={() => router.push('/privacy')}
-        style={[styles.button]}
-        variant="primary"
-      >
-        <Text style={styles.buttonText}>سياسة الخصوصية</Text>
+      <ThemedButton onPress={() => router.push('/privacy')} variant="primary">
+        <Text>سياسة الخصوصية</Text>
       </ThemedButton>
-      <ThemedButton
-        onPress={() => router.push('/contact')}
-        style={[styles.button]}
-        variant="primary"
-      >
-        <Text style={styles.buttonText}>تواصل معنا</Text>
+      <ThemedButton onPress={() => router.push('/contact')} variant="primary">
+        <Text>تواصل معنا</Text>
       </ThemedButton>
-      <ThemedButton
-        onPress={() => router.push('/about')}
-        style={[styles.button]}
-        variant="primary"
-      >
-        <Text style={styles.buttonText}>حول التطبيق</Text>
+      <ThemedButton onPress={() => router.push('/about')} variant="primary">
+        <Text>حول التطبيق</Text>
       </ThemedButton>
     </ThemedView>
   );
@@ -51,19 +32,5 @@ const styles = StyleSheet.create({
     gap: 20,
     justifyContent: 'center',
     alignItems: 'center',
-  },
-  button: {
-    width: '90%',
-    maxWidth: 640,
-    borderRadius: 5,
-    boxShadow: '0px 5px 5px rgba(0, 0, 0, 0.2)',
-    elevation: 5,
-  },
-  buttonText: {
-    //color: '#FFFFFF',
-    textAlign: 'center',
-    fontSize: 18,
-    fontWeight: '600',
-    paddingVertical: 10,
   },
 });